+++ packages/react-native/src/Link.tsx
@@ -1,37 +1,63 @@
 import React from "react";
 import { TouchableHighlight } from "react-native";
-import { useNavigationHandler } from "@curi/react-universal";
+import {
+  useNavigationHandler,
+  useStatefulNavigationHandler
+} from "@curi/react-universal";
 
 import { GestureResponderEvent } from "react-native";
 import { RouteLocation } from "@curi/router";
 import { NavigatingChildren } from "@curi/react-universal";
 import { NavType } from "@hickory/root";
 
-export interface LinkProps extends RouteLocation {
+export interface BaseLinkProps extends RouteLocation {
   onNav?: (e: GestureResponderEvent) => void;
   anchor?: React.ReactType;
-  children: NavigatingChildren | React.ReactNode;
   forward?: React.AnchorHTMLAttributes<HTMLAnchorElement>;
   method?: NavType;
+}
+
+export interface LinkProps extends BaseLinkProps {
+  children: React.ReactNode;
+}
+
+export interface NavLinkProps extends BaseLinkProps {
+  children: NavigatingChildren;
 }
 
 function canNavigate(event: GestureResponderEvent) {
   return !event.defaultPrevented;
 }
 
-const HookLink = React.forwardRef((props: LinkProps, ref: React.Ref<any>) => {
-  const { eventHandler, children } = useNavigationHandler<
-    GestureResponderEvent
-  >(props, canNavigate);
+export const Link = React.forwardRef(
+  (props: LinkProps, ref: React.Ref<any>) => {
+    const { eventHandler } = useNavigationHandler<GestureResponderEvent>(
+      props,
+      canNavigate
+    );
 
-  const { anchor: Anchor = TouchableHighlight, forward } = props;
+    const { anchor: Anchor = TouchableHighlight, forward, children } = props;
 
-  return (
-    // @ts-ignore
-    <Anchor onPress={eventHandler} ref={ref} {...forward}>
-      {children}
-    </Anchor>
-  );
-});
+    return (
+      <Anchor onPress={eventHandler} ref={ref} {...forward}>
+        {children}
+      </Anchor>
+    );
+  }
+);
 
-export default HookLink;
+export const NavLink = React.forwardRef(
+  (props: NavLinkProps, ref: React.Ref<any>) => {
+    const { eventHandler, navigating } = useStatefulNavigationHandler<
+      GestureResponderEvent
+    >(props, canNavigate);
+
+    const { anchor: Anchor = TouchableHighlight, forward, children } = props;
+
+    return (
+      <Anchor onPress={eventHandler} ref={ref} {...forward}>
+        {children(navigating)}
+      </Anchor>
+    );
+  }
+);