+++ packages/react-universal/src/hooks/useNavigationHandler.ts
@@ -7,8 +7,16 @@ import { NavType } from "@hickory/root";
 
 export type NavigatingChildren = (navigating: boolean) => ReactNode;
 
-export interface NavigationHookProps<T> extends RouteLocation {
-  children: NavigatingChildren | React.ReactNode;
+export interface NavigationHookProps<T> extends BaseNavigationHookProps<T> {
+  children: React.ReactNode;
+}
+
+export interface StatefulNavigationHookProps<T>
+  extends BaseNavigationHookProps<T> {
+  children: NavigatingChildren;
+}
+
+export interface BaseNavigationHookProps<T> extends RouteLocation {
   onNav?: (e: T) => void;
   forward?: object;
   method?: NavType;
@@ -20,15 +28,45 @@ function defaultCanNavigate() {
 
 export type CanNavigate<T> = (e: T, forward?: object) => boolean;
 
-export default function useNavigationHandler<
+export function useNavigationHandler<T extends React.BaseSyntheticEvent>(
+  props: NavigationHookProps<T>,
+  canNavigate: CanNavigate<T> = defaultCanNavigate
+) {
+  const { router } = useCuri();
+
+  function eventHandler(event: T) {
+    if (props.onNav) {
+      props.onNav(event);
+    }
+
+    if (canNavigate(event, props.forward)) {
+      event.preventDefault();
+
+      router.navigate({
+        method: props.method || "ANCHOR",
+        name: props.name,
+        params: props.params,
+        query: props.query,
+        state: props.state,
+        hash: props.hash
+      });
+    }
+  }
+  return {
+    eventHandler
+  };
+}
+
+export function useStatefulNavigationHandler<
   T extends React.BaseSyntheticEvent
 >(
-  props: NavigationHookProps<T>,
+  props: StatefulNavigationHookProps<T>,
   canNavigate: CanNavigate<T> = defaultCanNavigate
 ) {
   const { router } = useCuri();
   const cancel = React.useRef(undefined);
   const [navigating, setNavigating] = React.useState(false);
+
   React.useEffect(() => {
     return () => {
       if (cancel.current) {
@@ -45,15 +83,12 @@ export default function useNavigationHandler<
     if (canNavigate(event, props.forward)) {
       event.preventDefault();
 
-      let cancelled, finished;
       // only trigger re-renders when children uses state
-      if (typeof props.children === "function") {
-        cancelled = finished = () => {
-          cancel.current = undefined;
-          setNavigating(false);
-        };
-        setNavigating(true);
-      }
+      const done = () => {
+        cancel.current = undefined;
+        setNavigating(false);
+      };
+      setNavigating(true);
 
       cancel.current = router.navigate({
         method: props.method || "ANCHOR",
@@ -62,16 +97,14 @@ export default function useNavigationHandler<
         query: props.query,
         state: props.state,
         hash: props.hash,
-        cancelled,
-        finished
+        cancelled: done,
+        finished: done
       });
     }
   }
+
   return {
     eventHandler,
-    children:
-      typeof props.children === "function"
-        ? (props.children as NavigatingChildren)(navigating)
-        : props.children
+    navigating
   };
 }